python-mssql-fun/requirements.txt
Flask==1.1.2
pytest==6.1.2
flask-inputs==0.3.0
jsonschema==3.2.0
python-dotenv==0.15.0
pyodbc==4.0.30
flask_sqlalchemy==2.4.4
flask-cors==3.0.9
pandas==1.2.0
py-healthcheck==1.10.1
requests==2.27.1